+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 ### Refactoring
 - Removed deprecated database columns
 - Renamed project_favorites to project_preferences
+- Removed unused user stats page
 
 ## 0.29.0 (July 25, 2016)